Move tests to proper module and add total_at_height test
