+++ contracts/cw4-group/src/lib.rs
@@ -5,3 +5,6 @@ pub mod msg;
 pub mod state;
 
 pub use crate::error::ContractError;
+
+#[cfg(test)]
+mod tests;
